fix(runtime): route stale recovery through decisions
